requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
openai>=1.0.0,<2.0.0
httpx>=0.27.0,<1.0.0
opencv-python>=4.10.0.0,<5.0.0.0
numpy>=1.26.0,<2.0.0
tenacity>=8.2.0,<9.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.7.0,<3.0.0
loguru>=0.7.0,<1.0.0

